notebooks/[3]_rag_routing_and_query_construction_chinese.py
#!/usr/bin/env python3
# -*- coding: utf-8 -*-
"""
bRAG: 路由和查询构建
RAG Routing and Query Construction

本脚本演示了检索增强生成 (RAG) 系统中的路由和查询构建技术。
包括逻辑路由、语义路由、元数据过滤器、结构化搜索提示等高级功能。

作者: bRAGAI
版本: 1.0
"""

# ============================================================================
# 前置条件 (可选但推荐)
# ============================================================================

# 注意：如果您从未为此仓库创建过虚拟环境，请执行第一步。
# 否则，请确保您选择的 Python 内核来自您的 `venv/` 文件夹。

# 创建虚拟环境 (取消注释以执行)
# import subprocess
# subprocess.run(["python3", "-m", "venv", "../venv"])

# 激活虚拟 Python 环境 (取消注释以执行)
# subprocess.run(["source", "../venv/bin/activate"], shell=True)

# 如果您的 Python 不是来自您的 venv 路径，请确保您的 IDE 的内核选择
# (右上角) 设置为正确的路径 (您的路径输出应包含 "...venv/bin/python")

# 检查 Python 路径
import sys
print(f"当前 Python 路径: {sys.executable}")

# 安装所有包 (取消注释以执行)
# subprocess.run(["pip3", "install", "-r", "../requirements.txt", "--quiet"])

# 如果您选择跳过前置条件并仅使用全局 Python 路径环境安装此笔记本特定的包，
# 请执行下面的命令；否则，继续下一步。

# subprocess.run(["pip3", "install", "--quiet", "langchain_community", "tiktoken", 
#                "langchain-openai", "langchainhub", "chromadb", "langchain", 
#                "youtube-transcript-api", "pytube", "yt_dlp"])

# ============================================================================
# 环境设置
# ============================================================================

print("正在设置环境...")

# (1) 包导入
import os
from dotenv import load_dotenv

# 从 .env 文件加载所有环境变量
load_dotenv()

# 访问环境变量
langchain_tracing_v2 = os.getenv('LANGCHAIN_TRACING_V2')
langchain_endpoint = os.getenv('LANGCHAIN_ENDPOINT')
langchain_api_key = os.getenv('LANGCHAIN_API_KEY')

# LLM 模型
openai_api_key = os.getenv('OPENAI_API_KEY')

# (2) LangSmith 设置
# 参考文档: https://docs.smith.langchain.com/
os.environ['LANGCHAIN_TRACING_V2'] = langchain_tracing_v2
os.environ['LANGCHAIN_ENDPOINT'] = langchain_endpoint
os.environ['LANGCHAIN_API_KEY'] = langchain_api_key

# (3) API 密钥
print("环境设置完成！")

# ============================================================================
# 主要功能实现
# ============================================================================

def main():
    """
    主函数：演示 RAG 路由和查询构建功能
    """
    print("开始 RAG 路由和查询构建演示...")
    
    # 这里可以添加具体的 RAG 路由和查询构建实现代码
    # 例如：逻辑路由、语义路由、元数据过滤器、结构化搜索提示等
    
    print("RAG 路由和查询构建演示完成！")

if __name__ == "__main__":
    main()
