requirements_chinese.txt
# RAG实战 项目依赖包列表
# RAG Practice Project Dependencies

# Jupyter 内核支持
ipykernel

# RAG 相关包
ragatouille

# 环境变量管理
python-dotenv

# LangChain 核心包
langchain
langchain-community
langchain-core
langchain-openai

# 向量数据库客户端
pinecone-client

# 网页抓取和解析
beautifulsoup4

# 文本标记化
tiktoken

# 数值计算
numpy

# LangChain 模型和工具中心
langchainhub

# 本地向量数据库
chromadb

# YouTube 字幕 API
youtube-transcript-api

# YouTube 下载工具
pytube

# 文本重新排序 API
cohere

# PDF 处理
pypdf

# YouTube 下载工具 (替代方案)
yt_dlp
